refactored list view sorting into WithSortableDropdown trait
